+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VariableStarViewer/components/Controls/components/PeriodMultipleControls/theme.js
@@ -21,7 +21,7 @@ const theme = {
     border: {
       color: {
         light: 'light-6',
-        dark: 'light-6' // this will change when dark theme styles is added
+        dark: 'light-3'
       }
     },
     extend: () => css`
@@ -40,7 +40,7 @@ const theme = {
     hover: {
       border: {
         color: {
-          dark: 'accent-4',
+          dark: 'accent-2',
           light: 'neutral-2'
         }
       }